Fix ordering API: Refactored filters and updated tests
